complete content rel compute

+++ src/features/EventFeature.java
@@ -5,7 +5,7 @@ import source.Sentence;
 public class EventFeature implements FeatureCompute{
 
 	@Override
-	public float[] compute(Sentence sentence) {
+	public double[] compute(Sentence sentence) {
 
 		
 		return null;
