+++ src/features/FeatureCompute.java
@@ -4,6 +4,6 @@ import source.Sentence;
 
 public interface FeatureCompute {
 
-	public float[] compute(Sentence sentence);
+	public double[] compute(Sentence sentence);
 	
 }
